FunctionLiteral objects. They take the input parameters inside their scope.
We have yet to implement looking upwards in scopes to determine if a
the identifier is defined above. Thus, the factorial.ic example says
that there are two instances of undeclared identifiers. One because we
define factorial without declaring it, and one because the recursive
doesn’t see the identifier in the higher scope, even had it been
declared there.