Massive refactoring. Minor loss in LLVM output abilities, but road ahead looks much clearer
It became very frustrating to have Scope be an abstract class tied to
the AST. Instead, Scopes now form their own separate tree, and a
FunctionLiteral, for example) contains a scope. This gives scopes a
more concrete structure. They contain local and input parameters.
During the decl-finding and identifier-joining phases, all work output
is stored in a Scope. These patterns seem to be simpler than the
previous ones in that they trivially allow us to separate local and
input variables.