significantly more robust type dependency system
moved back on a feature (verifying return types in a function). the
reasoning for this is that it was part of verify_types(). however, this
can no longer be the case, as we may not know everything about the body
of the type at the time we go to do this verification. we must do this
and similar calls in another function called later.