most things are working. mutual_ptrs is still not.
mutual_ptrs fails because knowng the type of
f.x.y depends on knowing more than just that f.x has type &bar. because
we dereference it and then access the result, it actually depends on
knowing the type of @(f.x), but this isn’t encoded in any expression.
thus, the ordering chosen may put bar := type{…} after f.x.y, and the
compiler will fail. even with explicit @ this will fail.