Knee-deep in array refactoring. Probably will go chest-deep in the next few days.
Arrays were stored as a pointer to a buffer first containing a size and
then the data. The pointer pointed to the first element of the array so
that the size was behind it. We’re moving to a struct containing the
length and a pointer. This should help with both cache-locality and
potential alignment issues. This means we have to change both the code
that inits, prints, etc arrays as well as how we pass arrays around