better validation setup. still a long ways to go.
this is still a mess, but doing something right. need to figure out how
to be able to do analysis on cmds that don't have an associated
register. perhaps give them a register anyway with type void? since we
never need to store them we can give them numeric values that would
never be used (negatives?). sounds like register is starting to hold a
lot more information.