chainable operators now support operator overloading for structs
there's still an issue to decide about variants and arrays.

if i define (==) for a struct and have an array of structs, should it
find the global (==) as if equality on [--; T] was defined globally, or
should it find the local version as if there is a local definition of
equality for [--; T] at every local scope? or maybe a better way to
think of it is that when you define (==), you get the same on arrays
implicitly at the same scope whenever it is needed.