Resolving a parse ambiguity between scopes and functions.
The main issue is:
if (cond) { ... }
This could be parsed either assuming 'if' is a scope taking an argument, or as a
function returning a scope taking no arguments. In fact, depending on the
definition of 'if', either could be correct. Similar weird issues occur with
things like
if (a) == b { ... }
where we want (a) == b to be the argument to 'if', but instead we get 'if(a)'
treated as a function call.

A similar thing would happen with indexing and brackets: It could be an index or
an array literal argument.

By forcing scopes to have parentheses we avoid this problem because we can look
ahead to see if the close-paren is followed by an open brace. One subtlety here
is that if you prefer braces on new lines, you may be surprised to see this
treated as a function call, but we can probably give a good error message if we
see a function call on a scope followed by a line beginning with an open brace
(both illegal).

Ideally I'd like to allow args to scopes to not be parenthesized. While the
parser may not know the correct thing, there is actually only one correct parse
(though it's dependent on the types that we haven't yet computed). If we came up
with a sane way to handle this situation, that doesn't lead to bad things like
needing to write "typename" in certain places to coerce the parser to do the
right thing in C++, I would be pretty happy.

Also some parser clean-up.