Casting works with variants.
But there's probably something subtle I'm missing here. I think I'm working
under the assumption that the destination type is not a variant (the
expanded_size override field probably assumes this). In particular, what if I'm
casting a variant to itself? Probably instead of setting this to 1 I want to set
it to the expanded size of the destination type. Think on this some more.