Fixed parse ambiguity with functions and scopes.
Because scopes are expressions (which could evaluate to a type),
A -> B(C) { ... }
could be interpretted in two ways:
B(C) could be a type and this could be understood as a function literal of type
A -> B(C). Alternatively, B(C) { ... } could be a scope evaluating to a type T
and this could be understood as the type A -> T.

We resolve this ambiguity by saying that it means the former. If you want the
latter, you can explicitly request in with parentheses as in:
A -> (B(C) -> { ... })