Fixed bug printing a multi-dimensional array.
Printing arrays was done based on incrementing pointers, so if the data type had
size zero, it would only print the first element because we were already at the
end. Now we decrement based on the known size of the array. If the data members
are of size zero, we repeatedly load that element.

Note that there's room for optimization here because we know if the size of the
data element is zero or not at compile-time.