Refactoring
This seems to be helpful in reducing build times, but more importantly: whereas
previously we often wanted to work on a single phase (say, type verification) at
once over multiple AST nodes, it is more common now that we want to work on a
single type. The old approach was useful for getting started, but no longer
beneficial.