Lexing "block?" and parsing same as "block"
I'm opting for a slightly strange setup which I hope won't come back to bite me.
Using 'block' for both the type and the keyword means that '() -> block { ...
}' is ambiguous. It could be a function taking no arguments returning an object
of type 'block', or it could be the type of a function returning whatever 'block
{ ... }' evaluates to. The latter should always result in a type error because
block literals never evaluate to types, so we opt to parse it as the former. I'm
hesitant to say "never" but I'm confident enough to give this a shot.