Migrated to a better state for handling interfaces.
In the process, broke overloading/dynamic dispatch.
Essentially we now say uniformly that the actual parsed result is an
AST::Function node, regardless of whether it has const arguments or not. When we
go to call it, we use it to create a GeneratedFunction regardless of bound
arguments or not. This has some neat effects including that we now support
generic scope entry/exit.

However, for the moment declarations all have identical types of 'generic'. This
has been wrong forever anyways and needs to change but in particular now we need
to do a smart shadowing check.