Replace string with char_buffer which takes a length.
Not a great name, but the concept is better. It allows us to write strings on
our own. It can hold non-ascii encodings (which could be but are not yet part of
the type), and in this way differs from a character array. Also they should be
addable at compile-time, and += callable on them only at compile-time. The first
part is easy to implement today. The second needs a way to ensure a function is
only callable at compile-time.