EmItIR and EmitLVal alwasy return a vector.
This allows us to remove the recursiveness in IR::Val (which held a variant
containing a vector of Vals). This is probably a net loss for compilation speed,
but we can regain it in other ways. It does move us closer to very good options
for optimzing interpretter speed which is used in the compilation anyway, and
probably will end up as a net win on compilation anyway..