Structs are nominally typed now.
This leads to some oddities but I think is probably better in the long term.

For instance, struct {} != struct {}, but
s ::= struct {}
s == s

Returning a struct from a function like
f ::= () => struct {}
produces a unique type every time. This might be useful for tagged types or
generating strongly typed objects, but is certainly not obvious.

I wonder how useful it would be to use this to create unpronounceable types, so
that the only way to talk about the type of the object is to name the object
that has that type.