Checkpoint on generic structs work.
One of the problems we now is that generic structs can't just be called because
we need to know their arguments. In the example of

array ::= struct (T: type) {
  _size: int32
  _data: [*]T
}

The data pointer hasn't been verified yet. This means that we have to be careful
about what we can/cannot emit IR for. I believe in this case it's reasonable to
verify that the field types are in fact types. I'm hesitant to say this will be
a general solution.