Removing char from the language.
It's not that I don't want a character type, but this is the wrong abstraction
because it assumes an ascii character. I think we want a byte and codepoint
type. Strings shouldn't be arrays of characters but rather byte buffers with
tags determining if it's ascii or some other encoding.