Add a flag for loose casting numeric types.
This allows casting between integral types as well as between floating-point
types and from integral to floating even if it results in precision loss. This
enables testing a user-defined ascii character type which is just a `nat8`
wrapped in a struct. While such a type was always possible, it was previously
impossible to create a value of type `nat8`, so testing wasn't possible. Now we
can create ascii characters with  `char.asc(65 as nat8)`.

It's probably a good idea to have character literals, though I'm not entirely
sure what those would look like.