Adding argument-dependent lookup.
I'm sorry Titus. This may be naive, but because we have modules rather than
headers and namespaces this feels much safer. In particular, it's much less
likely you will pull in something you didn't intend to as exports are not
transitive. We also at very least get build breaks because there is
intentionally no way to discard function resolution candidates. If there is more
than one option, you'll hear about it rather than silently calling the wrong
one.

Probably long-term this won't stay so loose. I supsect we'll turn it off by
default and add #{adl} to enable it for specific overloads. Maybe operators will
get it by default and can #{!adl} or something. In any event we need this setup
or something similar because even if we selectively enable adl, we still need to
check what the selectively enabled candidates are and we only know this by
looking at the associated modules.

I also need to definitely turn it off for qualified calls, which I haven't done
yet.