Super hacky proof of concept for generic struct codegen.
Currently only works if there is exactly one generic struct in the
entire program, and it takes exactly one parameter and that parameter is
a type. And you only call it once.

That being said, it does cache appropriately so it can be self
referential.

I think next steps are to add IR operations for dealing with contexts
and adding bound constants to them. Though I'm worrying also about
copying bound constants from the context the struct literal was defined
in.