Get rid of repeated dispatch tables.
Instead, reuse a relevant node but tweak the address to be distinct. To
clarify `print a, b, c` needs 3 dispatch tables for each thing being
printed. We can't key them on the node for `a` because `a` itself might
be a function call needing a dispatch table of it's own. For example,
`print f(), g(), h()`. However, we can use the bottom bits of the
pointer to indicate that his is not for `f()` but for some expression
containing `f()` (of which there must only be one).