Fixing some bugs in the block-combining optimization pass.
Aside from just being incorrect, we also needed to make sure all other
references to blocks were handled appropriately. This means any phi node
on a block with one predecessor needed to be fixed. The number of
predecessors on a block with a phi-node must be the number of entries in
the node itself, so we just need to either fix all phi nodes with one
entry or ensure we never create any in the first place. We opted for the
second solution.