Fixing a bug in `Context`-switching between modules during type verification.
I still don't totally know what the rules surrounding this should be. In
particular, when I should switch and how to know when to switch back.
Moreover, the benefit is that today each module is compiled on it's own
thread, but that may not always be the case. Modules compilation may
benefit from separating out top-level functions each on their own thread
for instance so you still need more synchronization. Maybe this should
all be one global structure, but then I probably want a mechanism to
avoid rehashing the whole thing. Maybe next steps are to make it one
global hash map to avoid the context switch and deal with perf
consequences as they arise.