Removing interfaces.
I want *something* like this but this version was never fully fleshed out, and
the correct thing is hard/I don't know what it is. In particular, interfaces
like this can never be changed. With enough uses, in aggregate callers will not
enable you to add any requirements (some caller won't satisfy it). Similarly,
with enough uses the callees will require all requirements. Things that can't
change are dangerous.