More optimization
Rather than having an untyped buffer with packed data for the registers,
we just align all the data at some width wide enough to hold any type.
We might want to write to a register.

This has significant performance improvements as we don't need to do a
hash lookup for each access. Moreover, We don't need to construct the
map or store it during code emission.

For counting primes < 2000000, we now execute in ~1.7 seconds.