Fixing function calls.
This was a bad bug, and I'll likely see it again. I was passing an
unaligned_ref<T> into untyped_buffer::set which deduced the type as the
unaligned_ref rather than decaying as I would have liked. I was able to
debug it by looking at the actual value written and seeing that it was
pointer like (i.e., started 0x00007f...). In fact it shared the first
large number of bytes in common with other things in the untyped_buffer,
which should have been a bigger clue. that it was referencing something
nearby. Unfortunately I printed the type from the nearby
using-declaration  while debugging rather than decltype(<offending
expr>).