Migrating back to separate instructions.
I made the mistake of thinking that commands each being their own type
was slow to execute and therefore we need to pack them in the buffer.
This was true: They were slow to execute, but that doesn't mean they
need to be packed from the start. We want separate instruction types for
the purposes of optimization passes, but before we execute, we want to
serialize them to a buffer.

There is going to be some tradeoff based on the size of functions. It's
probably not worth serializing really small functions. Like the ones
constructed just to evaluate a type. We'll see.