Simplified inliner.
Needed to separate out instructions from the base class and have basic
blocks only depend on the base class. The base class does need to
forward-declare the inliner. This is all very sad, but I don't know how
to fix it. I tried having a giant variant of all instructions, but it's
200+ variant members which GCC didn't like very much.