Move the more common instructions from a hash map to an array.
We're not done with this migration, but it accounts for a good
performance win. On our prime number benchmark, we are now ~16% faster
than Python. There are more earnings to be had as we haven't migrated
the load instruction.

It's also noteworthy that it is still significantly more expensive to
call a function in our interpretter than it is in Icarus. We haven't
really spent any time optimizing that codepath, which is also extremely
important.