Keep all the sources alive.
This is not the end-state we want but it's an okay intermediate. What we
really want is a way to open/close sources so we don't hold on to their
memory, but can reopen them later if we need to emit diagnostics. This
requires some amount of central management for which we now have the
bare bones, but presumably it'll need to be upgraded/rewritten later.
This just unblocks showing source quotes in the diagnostics again.