Work on fixing incomplete structs.
The comment added in explains one issue fix. The other is the constant
slot reservation mechanism (not yet fixed). Constants can reserve slots
even when they're incomplete. And later uses of the declaration will
early-return because they see a slot.

We could fix this by having slots know if they are complete or not. This
feels a lot like the completeness bool we added to structs. Perhaps we
don't need both.