Removing variants and dynamic dispatch.
Variants are not represented in the format I want them, nor is the
code-gen what I want. It also complicates generic functions and overload
resolution. Because I'm focussing on generics for the time being, it'd
be nice to simplify the code there first before adding in dynamic
dispatch the way I actually want it.