A better way to handle destructors.
This still compiles the destructor before the type is complete which is
problematic, but it gives a good idea about how to proceed. During body
verification, aggregate a list of all things that need to happen before
the body for this function can be emit and then iterate through those in
some topologically sorted order.