Fixes #21
Calling across module boundaries was broken because we did not have the
function's type from the imported module tied to the declaration in the
importing module. This meant during overload resolution, calls across
module boundaries would fail to find the type and crash during
type-checking.

This is fixed by adding the qualified type to the imported declaration
on the importing module, even though the declaration is not made inside
that module. We used to have the invariant that any expression whose
qualified type was set on a module was necessarily part of the syntax
tree for that module. This was never explicitly stated/guaranteed or
luckily relied upon. It is no longer true.