Fix #20. Separate out copy/move assignment.
Both have signature `(*T, *U) -> ()`. Binding to `(move)` or `(copy)`
makes it a move or copy assignment respectively.

If a type has no assignment operators, copy and move assignments will be
generated for each by pointwise moving/copying the struct members.

If a type has a copy- but no move-assignment from a given type, the copy
will be used in place of the move if a move-assignment from that type is
requested.