Fix #59
The issue was that getline was writing to an address assuming a `char *`
was stored there, but the interpreter had used another type `ir::Addr`
which was not layout compatible.

This means that if we want foreign functions to interact nicely with the
interpreter, the interpreter needs to use `char *` for pointers.

This change does away with `ir::Addr`, replacing it with `ir::addr_t` as
an alias for `char *`.