Rewrite of scope building.
The approach here isn't complete but it's much simpler to understand. We
effectively do a breadth-first traversal of the jump blocks. If we ever
get to a choose-jump or a jump exit, we inline the appropriate arguments
and resolve the appropriate blocks.

We have removed the capability to have before functions for the time
being, but this approach makes it easier to see how that gets
implemented, and also easier to understand how we will end up emitting
destructors and dealing with (labelled) yields.