Fixing some cross module issues, but not in a robust way.
The problem is that we incorrectly assume that either we are a context
in the tree of the primary module being processed, or we are on the root
of an imported module. This was because only roots can be exported from
imported modules. This is true of module scopes, but instantiating a
generic across module boundaries will still end up on the root scope but
not the root context. We need a way to access that, and the current
approach doesn't make that easy.