Relying exclusively on the WorkGraph system.
Primarily the change here is to establish a global set of work that we
add to and mark as complete. WorkGraph's then take work from the set and
complete them. This way instantiations of multiple work graphs do not
end up completing each other's work and we don't end up with infinite
loops evaluating cyclic dependencies. This happened easily before with
self-referential structures.

Amazingly, we removed a bunch of cruft like `must_complete` and some
data stored on `compiler::Context`.