Improving compiler work ordering.
Not a great approach but this does seem to do the job. Definitely worth
revisiting though.

We reverted the previous separation for designated initializers because
we're not going the coroutine-y route and it's no longer necessary or
useful.

Fixes #76 -- The issue wasn't that the lookup was incorrect, but rather
that we weren't verifying the type ahead of time. By moving the code to
a separate module, we ensured completion of all type verification steps,
resulting in the observed behavior.