Fix #87
When a module is imported, we attempt several different paths, some of
which may fail and set errno. We were never cleaning this up. This is a
subtle problem because we never want errno being set from compilation to
affect the interpreter or compile-time execution, even though
interpreted code will use the same instance of errno as the compiler
itself. We're fixing this by saving and resetting errno on each
interpreter invocation.