Add user-defined error message expresison.
Fixes #66.

This is mostly useless right now. If the expression is ever compiled it
will generate an error. The only way it could be not-compiled is if it
is in the body of a generic, but we don't have a mechanism to have it be
conditionally evaluated. This means that the generic couldn't be
instantiated at all.

Longer-term, when we have compile-time if-statements that only do
type-checking on the taken branch, this tool will be invaluable.