Adding a "has_block" builtin function.
has_block takes a `scope_context` and a `[]char`, returning whether or
not there is a block by that name in the scope.

This fixes #100, given that the scope author properly checks for
optional scopes.

This isn't a long-term solution, but is a reasonable stop-gap, and
something like this may persist.