ProcessModule was a mistake.
We need to fully process modules that are imported. We could maybe get
away with only doing type verification in some circumstances, but it's
better to handle the full thing.

To that end, starting to handle some things that were type-checked but
had no code emission implementation. In particular, designated
initialization with an expression initializing multiple fields.