Stop storing subroutine in ir::Module.
This brings about a regression for builtin.debug_ir which needed access
to that for printing purposes. One solution would be to move that off
the builtin module and make it a truly magical builtin so that during
compilation we can see that it's called and have the calling scope save
this string directly into the debug instruction.