Basic infrastructure required for module imports.
Relying on import being a member of builtin for the moment, but this
will not last. It's nicer syntactically to have import be a unary prefix
operator, and using a builtin has weird consequences. It would be the only
builtin whose evaluation could effect the state of compilation, and it
would be difficult to short-circuit execution without support from
jasmin.