Begin separating type categories into their own targets.
The primary value of this is that IrFunction no longer needs to depend
on a target containing all types, so dependent types can depend on it.

I don't think this model will enable us to fully have separate targets
(e.g., Type is contsructible from PointerType), so it's not fully
decoupled, but it at least provides improved type-safety.